BUG: Several LLVM runtime functions returned 32-bit values for 64-bit functions.  So LLVM was calling a function that returned 32-bit values in a 64-bit register, resulting in chaos with optimizations on.  BUG: Evidently GCC aggressively optimizes bools that hold values other than true/false, switched to unsigned char to get the right behavior for some tests.
Former-commit-id: 6041e83b7de7d4c8e827bf05a3e761898e3f952f

+++ branches/ocelot-ptx-2.1/ocelot/executive/implementation/LLVMRuntimeLibrary.inl
@@ -139,7 +139,7 @@ extern "C"
 		return hydrazine::bitFieldInsert( in, orig, position, length );
 	}
 
-	unsigned int __ocelot_bfi_b64( long long unsigned int in, 
+	long long unsigned int __ocelot_bfi_b64( long long unsigned int in, 
 		long long unsigned int orig, unsigned int position, 
 		unsigned int length )
 	{
@@ -151,7 +151,8 @@ extern "C"
 		return hydrazine::bfind( a, shift );
 	}
 
-	unsigned int __ocelot_bfind_b64( long long unsigned int a, bool shift )
+	long long unsigned int __ocelot_bfind_b64( long long unsigned int a,
+		bool shift )
 	{
 		return hydrazine::bfind( a, shift );
 	}
